+++ ingest2/operations/stub_op.go
@@ -9,7 +9,7 @@ type stubOpHandler struct {
 	effectsProvider
 }
 
-//Details - used as temporary solution for not handled operations
+// Details - used as temporary solution for not handled operations
 func (h *stubOpHandler) Details(op rawOperation, opRes xdr.OperationResultTr) (history2.OperationDetails, error) {
 	return history2.OperationDetails{
 		Type: op.Body.Type,