+++ CHANGELOG.md
@@ -1,4 +1,31 @@
-# Unreleased
+# 3.0.1-x.7
+
+### Added 
+
+* Signers endpoint
+* Account/Signer Role/Rules
+* Endpoints for reviewable requests
+* Docs on reviewable requests and sale
+
+### Changed
+
+* `Details`/`Reason`/`NewDetails` field names in reviewable request types unified to `creatorDetails`
+* Corresponding messages in `messages` map
+* Calculated fee response
+
+### Removed
+
+* `UpdateSaleEndTimeRequest` type
+
+### Fixed
+
+* Fee bounds
+* Fee calculation for account
+
+# 3.0.1-x.6
+
+### Fixed
+* 500 on fees
 
 # 3.0.1-x.5
 
@@ -11,6 +38,7 @@
 * 500 error on /history endpoint when receiving history of payments
 * Fixed issue with balanceID not been set for balance effects
 * Fixed Fee 500 if asset does not exists
+
 
 # 3.0.1-x.4
 
@@ -63,7 +91,6 @@ Added back proxy to API
 
 ## Added
 
-* `Limits` and `ExternalSystemIDs` to `/accounts` endpoint
 * `/sales` endpoint
 * `/order_book` endpoint